python-showcase/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
kafka-python==2.0.2
google-cloud-storage==2.16.0
google-cloud-bigquery==3.17.2
google-cloud-pubsub==2.20.2
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0
faker==24.2.0
joblib==1.3.2
requests==2.31.0
